Warning suppressed on IA84 arch.
git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@14025 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/schedulers/sched_opt/src/nm_so_pkt_wrap.c
@@ -394,7 +394,7 @@ nm_so_pw_finalize(struct nm_so_pkt_wrap *p_so_pw)
     if(!remaining_bytes) {
       /* Headers area continues on a separate iovec entry */
       vec++;
-      ptr = (void *)nm_so_aligned((uint64_t)vec->iov_base);
+      ptr = (void *)nm_so_aligned((intptr_t)vec->iov_base);
       remaining_bytes = vec->iov_len
 	                - (ptr - vec->iov_base);
       to_skip = 0;
